Migrate ScheduleGroupScoreUpdatesJob
